pyproject.toml
[project]
name = "a2a-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk[http-server]>=0.3.25",
    "openai>=2.30.0",
    "uvicorn>=0.30.0",
]
